Addition of Mnemonic Parsing, Vbox, openFile, and editMenu functions
All of the menus now have the mnemonic parsing on the letters that each
should have, Vbox was added for the menu bar for better stacking got
the borderline, openFile was modified to set text but cannot currently
get dimensions back. editMenu had functionality added to them for cut,
paste, copy and delete.